refactor: 승무패 열거형 클래스 네이밍을 GameResult, 결과 계산 클래스 네이밍을 GameResultRecord으로 수정
