+++ src/main/kotlin/blackjack/domain/model/GameResult.kt
@@ -1,0 +1,7 @@
+package blackjack.domain.model
+
+enum class GameResult {
+    WIN,
+    DRAW,
+    LOSE,
+}
