--- src/main/kotlin/blackjack/domain/model/WinLoss.kt
@@ -1,9 +1,0 @@
-package blackjack.domain.model
-
-enum class WinLoss(
-    val koreanText: String,
-) {
-    WIN("승"),
-    DRAW("무"),
-    LOSE("패"),
-}
